sniffer/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
